+++ src/tree-provider/mpq/mpq-item-node.ts
@@ -14,6 +14,9 @@ export class MpqItemNode extends vscode.TreeItem {
         this.contextValue = 'folder';
         if (this.node.isLeaf()) {
             this.contextValue = 'file';
+            if (label.toLowerCase().endsWith('.mdx')) {
+                this.contextValue = 'mdxFile';
+            }
             this.command = command || {
                 title: 'Open with mpq',
                 command: 'blpPreview.openMpq',